Factorise out NFC and External Storage modules.
